Delete car with car.ownerId

+++ frontend/Table-Main-Master-main/src/app/shared/component/car-delete/car-delete.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { CarService } from 'src/app/core/service/car.service';
 import { SessionStorageService } from 'src/app/core/service/session-storage.service';
 import { DeleteModalService } from 'src/app/core/service/delete-modal.service';
+import {Car} from "../../../core/Interface/car";
 
 @Component({
   selector: 'app-car-delete',
@@ -12,6 +13,9 @@ import { DeleteModalService } from 'src/app/core/service/delete-modal.service';
 export class CarDeleteComponent {
   carId: number;
   isOpen: boolean;
+  ownerId: number;
+  car: Car;
+  carOwnerId: number;
 
   constructor(private carService: CarService,
     private sessionStorageService: SessionStorageService,
@@ -19,24 +23,40 @@ export class CarDeleteComponent {
     public deleteModalService: DeleteModalService) { }
 
   ngOnInit(): void {
+      this.ownerId = this.sessionStorageService.getItem('id');
     this.carId = this.sessionStorageService.getItem('carId');
+    console.log(this.carId)
     this.deleteModalService.isOpen$.subscribe(isOpen => {
       this.isOpen = isOpen;
     });
+
+    this.carService.getCarById(this.carId).subscribe((car: Car) => {
+      this.car = car;
+      this.carOwnerId = car.ownerId
+      console.log("oID",this.carOwnerId);
+    });
+
+
   }
 
   deleteItem() {
-    this.carService.deleteCar(this.carId).subscribe(
-      () => {
-        console.log("Car successfully deleted");
-        this.deleteModalService.closeModal();
-        this.router.navigate(['/']);
-      },
-      (error) => {
-        console.error("Error deleting car:", error);
-        this.deleteModalService.closeModal();
-      }
-    );
+
+   if(this.ownerId == this.carOwnerId){
+     console.log("proslo")
+     this.carService.deleteCar(this.carId).subscribe(
+       () => {
+         console.log("Car successfully deleted");
+         this.deleteModalService.closeModal();
+         this.router.navigate(['/']);
+       },
+       (error) => {
+         console.error("Error deleting car:", error);
+         this.deleteModalService.closeModal();
+       }
+     );
+   } else {
+     alert("Ne mozete izbrisati oglas drugog ownera")
+   }
   }
 
   openModal() {
